[ASPA-027] Made the to-be-disabled buttons look like they are disabled
Issue: The look of the disabled buttons needed to be changed.
Solution: A class 'btn-disabled' as added to the classList of the respective buttons. 'btn-disabled' was defined in the aspa.webflow.css file. Also the opacity of the buttons was reduced. The buttons were also made visually unclickable.
Risk: None expected
Reviewed by: Raymond Feng