[ASPA-027] Made the to-be-disabled buttons look and behave like they are disabled (#29)
* [ASPA-027] Corrected the parameter type in the comment

* [ASPA-027] Stopped the button click event listeners from being added to the to-be-disabled buttons

* [ASPA-027] Made the to-be-disabled buttons look like they are disabled

Issue: The look of the disabled buttons needed to be changed.
Solution: A class 'btn-disabled' as added to the classList of the respective buttons. 'btn-disabled' was defined in the aspa.webflow.css file. Also the opacity of the buttons was reduced. The buttons were also made visually unclickable.
Risk: None expected
Reviewed by: Raymond Feng

* [ASPA-027] Removed a debugging line of code

* [ASPA-027] Undid the erroneously deleted line in previous commit and deleted the correct debugging line