[ASPA-035] Directly accessing url ErrorWrapping (#38)
Issue:
Directly accessing some of the controller methods will throw a PHP error that is visible to the users. The problem is due to missing GET/POST variables or the logic should not allow direct external access. This issue exists for Stripe Payment successful page and therefore it may also be a problem for all other controller methods. 

Solution:
Wrapping a CodeIgniter error if it's missing parameters or the logic appears that the user is directly accessing it.
A wiki page is also made to specify the Error Codes used in the Error pages. https://github.com/UoaWDCC/ASPA-EnrollmentForm/wiki/Back-End-API-Error-Codes
* EnrollmentForm/MakeStripePayment
* EnrollmentForm/send_email (REMOVED)
* EnrollmentForm/validate
* EnrollmentForm/StripePaymentSuccessful
* EnrollmentForm/IEPayPaymentSuccessful (REMOVED)
* EnrollmentForm/LoadOfflinePayment

Risk:
Mypay's payment function (MakeMypayPayment) might have the same problem and needs to be checked.
The send_email controller function has been removed.

Reviewed By:
Raymond