[ASPA-034] Made the ok button functioning when the information on the page has not changed
Issue: The okay button was not working when the user navigates from page 3 to page 4 and then navigates back to page 3, when the data in the name and email fields has not changed.
Solution: The ok button was disabled so the user cannot make repeated requests while on the page. The ok button only gets reenabled when the user navigates to a different page or when an error is identified on the same page. Also, the previous implementation of preventing repeated requests was removed.
Risk: Nothing
Reviewed by:  Raymond Feng, Lucas Gao