[ASPA-034] Made the ok button functioning when the information on page 3 remains unchanged (#39)
* [ASPA-034] Made the ok button functioning when the information on the page has not changed

Issue: The okay button was not working when the user navigates from page 3 to page 4 and then navigates back to page 3, when the data in the name and email fields has not changed.
Solution: The ok button was disabled so the user cannot make repeated requests while on the page. The ok button only gets reenabled when the user navigates to a different page or when an error is identified on the same page. Also, the previous implementation of preventing repeated requests was removed.
Risk: Nothing
Reviewed by:  Raymond Feng, Lucas Gao

* [ASPA-034] Just improved a grammar issue

* [ASPA-034] Fixed the repeated requests problem due to enter key

Issue: The enter key could be pressed even when the ok button is visually disabled, leading to repeated requests if the user presses 'enter' multiple times
Solution: Inserted a line of code in the key bindings code chunk which checks if the ok button on page 3 is disabled before making the enter key have its effect
Risk: Nothing

* [ASPA-034] Changed if statement function alignment

Purely cosmetic change to code base.

Co-authored-by: Raymond Feng <raymondf.5522@gmail.com>