[ASPA-016] Bug fix empty entry (#51)
Issue:
Trying to register only inputting an email address (no name in the field)

Solution:
Change the logic for button clicking. Checked the name and the email is empty separately using two if-else statements. And another if statement for checking if either is empty.

Risk:
None for now

Reviewed By:
James You