[ASPA-9] Extract name and UPI from membership spread
Issue:
Name field not necessary, only email is needed. We can get user's Name, UPI and UID from the email.

Solution:
Check the input email against the ASPA's membership spreadsheet, and if it matches a user, we will pull the name (and the UPI if it exists) from the spreadsheet, and add this information into each created sheet (right next to each other).
Remove the name field and all the javascript code related to the name field

Risk:
For the dev environment, it won't update the upi if you already on the spreadsheet, so if you want to test it please delete your entry on the dev registration spreadsheet :)

EDIT (from @Ray-F): Some people's names have an apostrophe in the them, and this causes issues when sending emails. We need to make sure this is released into master at the same time as @Metarock (John)'s issue of ASPA-35.

Reviewed By:
April, James, Raymond